Reformat display of ingredients
frontend has a fixed number of ingredients that are updated one by one.
If a recipe has fewer ingredients, the remaining ones are set to an empty
string. If a recipe has more ingredients, an error should be thrown.
This is less than ideal and should be replaced by a solution that allows
arbitrary numbers of ingredients.